cleanup test_coarse_tile_e2e: update skips and test comments (#4175)
* test: un-skip passing coarse_tile_e2e tests, refresh stale skip reasons

Now that PR #4133 (B+H coarse-tiling fixes) and PR #4140 (copy_forced
DCE fix) are both on main, empirically re-ran every skipped/xfailed
test in test_coarse_tile_e2e.py to check whether its skip is still
accurate.

Removes skips for 9 tests that now pass outright:
test_flash_tile_H_Lq, test_flash_tile_Lq, test_flash_v2_tile_Lq,
test_flash_v2_tile_H_Lq, test_flash_v3_tile_Lq, test_flash_v3_tile_B_H,
test_flash_v3_tile_H_Lq, test_hint_flash_attention_v2, and
test_hint_flash_attention_v3.

Rewords skip reasons for tests that still fail, but no longer with the
symptom their reason describes (mostly issue #4126's "copy_forced
drop -> inf" framing, which is now a large-but-finite mismatch, or a
completely different error) so the text reflects the currently
observed failure rather than a historical one.


* test: drop historical failure narrative from test_hint_flash_attention_v2

Now that the test passes (previous commit), its docstring no longer
needs to describe the restickify-infeasibility failure it used to hit
or the PR/CI history behind the decision to xfail it.


* test: move hbm_pool e2e tests from test_coarse_tile_e2e.py

Relocate the five hbm_pool/pool_size threading tests
(test_bundle_pool_size_threaded_from_hbm_pool_sizes,
test_pool_alloc_scoped_per_bundle_across_fallback_boundary,
test_no_python_side_pool_tensor_allocated,
test_pool_size_kwarg_in_generated_sdsc_call,
test_pool_size_kwarg_omitted_when_no_pool) into a new
TestHbmPoolPlanningE2E class in test_hbm_pool_planning.py, alongside the
existing hbm_pool_planning unit tests, rather than in the coarse-tile e2e
file.

Fix a latent bug the move exposed: two tests asserted assertNotIn("_pool_",
src) to check for the absence of a Python-side pool tensor, but this
file's own name (test_hbm_pool_planning.py) contains "_pool_" and leaks
into generated source via embedded SourceLoc debug-handle file paths.
Replaced with a regex matching the actual _pool_{kernel_name} = ...
assignment pattern emitted by call_kernel().


---------
