+++ tests/inductor/test_coarse_tile_e2e.py
@@ -2405,7 +2405,10 @@ def test_outside_consumer_two_accum_512x256_B4():
 
 
 @pytest.mark.skip(
-    reason="infeasible restickify for 1D denom in mixed 1D/2D tiled scope"
+    reason="Unsupported: partial reduction result consumed before"
+    " accumulation is complete (op reads a per-tile partial result from"
+    " the same loop group) -- same carry-propagation gap as the Lk"
+    " reduction-dim tiling cases below"
 )
 def test_outside_consumer_two_accum_512x256_A4_B4():
     """Flash-style: out=zeros, denom=zeros; tiled copy_; return out/denom — A÷4 B÷4."""
@@ -2867,21 +2870,6 @@ def test_flash_tile_B():
     )
 
 
-@pytest.mark.skip(
-    reason=(
-        "Compiles now (the squeeze-position crash from issue #3613 is "
-        "fixed), but produces numerically wrong results: ~93% of output "
-        "elements mismatched, spread across both Lq tiles and all H heads. "
-        "Root cause not yet isolated -- ruled out so far: the "
-        "_tiled_dims_for_dep raw->squeezed fix (removing it causes an "
-        "immediate validate_writer_tile_advance failure, so it's necessary "
-        "and unrelated), the _insert_one_read_copy active_full_sizes fix "
-        "(reverting it does not change the mismatch), and the loop_internal "
-        "output_tiled_dims-clearing for op8/op9 (un-clearing it does not "
-        "change the mismatch either). See issue #3613 for the ongoing "
-        "investigation."
-    )
-)
 def test_flash_tile_Lq():
     """Flash v1: tile Lq÷2 only."""
     run_coarse_tile_test(
@@ -2919,15 +2907,6 @@ def test_flash_tile_B_H():
     )
 
 
-@pytest.mark.skip(
-    reason=(
-        "validate_writer_tile_advance now catches this at compile time: "
-        "squeeze-position bug in _insert_copy_op's write-side "
-        "_tiled_dims_for_dep (raw d{N} numbering breaks when a unit dim is "
-        "squeezed out of the index). Same root cause as issue #3613; "
-        "deferred until PR #3622's tile.py helpers land."
-    )
-)
 def test_flash_tile_H_Lq():
     """Flash v1: tile H÷4 Lq÷2."""
     run_coarse_tile_test(
@@ -3096,8 +3075,9 @@ def _flash_v2_fn(
 
 
 @pytest.mark.skip(
-    reason="copy_forced(running_max, real_max) write-back is dropped -- real_max"
-    " never carries across H tiles, producing inf (issue #4126)"
+    reason="H-tiling produces a large finite mismatch (~88% of elements,"
+    " max abs diff ~4) -- distinct, still-open H-tiling bug, not"
+    " accumulator inf"
 )
 def test_flash_v2_tile_H():
     """Flash v2: tile H÷4 only."""
@@ -3113,9 +3093,9 @@ def test_flash_v2_tile_H():
 
 
 @pytest.mark.skip(
-    reason="B-tiling still produces inf after the copy_forced/"
-    "mutation_write_back drop fix (issue #4126) -- distinct, still-open"
-    " B-tiling bug, not the copy_forced accumulator drop"
+    reason="B-tiling produces a large finite mismatch (~48% of elements,"
+    " max abs diff ~3.5) -- distinct, still-open B-tiling bug, not"
+    " accumulator inf"
 )
 def test_flash_v2_tile_B():
     """Flash v2: tile B÷2 only. B=2."""
@@ -3128,10 +3108,6 @@ def test_flash_v2_tile_B():
     )
 
 
-@pytest.mark.skip(
-    reason="copy_forced(running_max, real_max) write-back is dropped -- real_max"
-    " never carries across tiles, producing inf (issue #4126)"
-)
 def test_flash_v2_tile_Lq():
     """Flash v2: tile Lq÷2 only."""
     run_coarse_tile_test(
@@ -3159,9 +3135,9 @@ def test_flash_v2_tile_Lk():
 
 
 @pytest.mark.skip(
-    reason="B-tiling still produces inf after the copy_forced/"
-    "mutation_write_back drop fix (issue #4126) -- distinct, still-open"
-    " B-tiling bug, not the copy_forced accumulator drop"
+    reason="B/H-tiling produces a large finite mismatch (~46% of elements,"
+    " max abs diff ~3.8) -- distinct, still-open B-tiling bug, not"
+    " accumulator inf"
 )
 def test_flash_v2_tile_B_H():
     """Flash v2: tile B÷2 H÷4. B=2."""
@@ -3174,9 +3150,6 @@ def test_flash_v2_tile_B_H():
     )
 
 
-@pytest.mark.skip(
-    reason="finalize_layouts: restickify infeasible for copy ops across loop groups"
-)
 def test_flash_v2_tile_H_Lq():
     """Flash v2: tile H÷4 Lq÷2. Equivalent to original test_flash_v2."""
     run_coarse_tile_test(
@@ -3351,9 +3324,9 @@ def test_flash_v3_tile_H():
 
 
 @pytest.mark.skip(
-    reason="B-tiling still produces inf after the copy_forced/"
-    "mutation_write_back drop fix (issue #4126) -- distinct, still-open"
-    " B-tiling bug, not the copy_forced accumulator drop"
+    reason="B-tiling produces a large finite mismatch (~48% of elements,"
+    " max abs diff ~3.5) -- distinct, still-open B-tiling bug, not"
+    " accumulator inf"
 )
 def test_flash_v3_tile_B():
     """Flash v3: tile B÷2 only. B=2."""
@@ -3366,15 +3339,6 @@ def test_flash_v3_tile_B():
     )
 
 
-@pytest.mark.skip(
-    reason=(
-        "validate_writer_tile_advance now catches this at compile time: "
-        "squeeze-position bug in _insert_copy_op's write-side "
-        "_tiled_dims_for_dep (raw d{N} numbering breaks when a unit dim is "
-        "squeezed out of the index). Same root cause as issue #3613; "
-        "deferred until PR #3622's tile.py helpers land."
-    )
-)
 def test_flash_v3_tile_Lq():
     """Flash v3: tile Lq÷2 only."""
     run_coarse_tile_test(
@@ -3400,11 +3364,6 @@ def test_flash_v3_tile_Lk():
     )
 
 
-@pytest.mark.skip(
-    reason="B-tiling still produces inf after the copy_forced/"
-    "mutation_write_back drop fix (issue #4126) -- distinct, still-open"
-    " B-tiling bug, not the copy_forced accumulator drop"
-)
 def test_flash_v3_tile_B_H():
     """Flash v3: tile B÷2 H÷4. B=2."""
     run_coarse_tile_test(
@@ -3416,15 +3375,6 @@ def test_flash_v3_tile_B_H():
     )
 
 
-@pytest.mark.skip(
-    reason=(
-        "validate_writer_tile_advance now catches this at compile time: "
-        "squeeze-position bug in _insert_copy_op's write-side "
-        "_tiled_dims_for_dep (raw d{N} numbering breaks when a unit dim is "
-        "squeezed out of the index). Same root cause as issue #3613; "
-        "deferred until PR #3622's tile.py helpers land."
-    )
-)
 def test_flash_v3_tile_H_Lq():
     """Flash v3: tile H÷4 Lq÷2. Equivalent to original test_flash_v3 (small sizes)."""
     run_coarse_tile_test(
@@ -3567,7 +3517,9 @@ def _flash_v4_fn(q, k, v, *, B, S, H, D, b_tiles=1, h_tiles=1, lq_tiles=1, lk_ti
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="AssertionError in _stick_symbol: within-stick coordinate"
+    " carries 2 free symbols, want exactly 1 -- v4's view+transpose"
+    " layout produces a coordinate expression padding.py cannot solve"
 )
 def test_flash_v4_tile_H():
     """Flash v4: tile num_heads÷4 only."""
@@ -3579,7 +3531,9 @@ def test_flash_v4_tile_H():
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="AssertionError in _stick_symbol: within-stick coordinate"
+    " carries 2 free symbols, want exactly 1 -- v4's view+transpose"
+    " layout produces a coordinate expression padding.py cannot solve"
 )
 def test_flash_v4_tile_B():
     """Flash v4: tile batch_size÷2 only. B=2."""
@@ -3591,7 +3545,9 @@ def test_flash_v4_tile_B():
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="AssertionError in _stick_symbol: within-stick coordinate"
+    " carries 2 free symbols, want exactly 1 -- v4's view+transpose"
+    " layout produces a coordinate expression padding.py cannot solve"
 )
 def test_flash_v4_tile_Lq():
     """Flash v4: tile max_seqlen_q÷2 only."""
@@ -3603,7 +3559,9 @@ def test_flash_v4_tile_Lq():
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="AssertionError in _stick_symbol: within-stick coordinate"
+    " carries 2 free symbols, want exactly 1 -- v4's view+transpose"
+    " layout produces a coordinate expression padding.py cannot solve"
 )
 def test_flash_v4_tile_H_Lq():
     """Flash v4: tile num_heads÷4 max_seqlen_q÷2. Equivalent to original test_flash_v4."""
@@ -3617,7 +3575,9 @@ def test_flash_v4_tile_H_Lq():
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="Hangs rather than failing fast (observed: no completion after"
+    " 2+ minutes, killed) -- distinct from the other v4 tile combinations,"
+    " which fail at compile time; root cause not yet investigated"
 )
 def test_flash_v4_tile_H_Lq_Lk():
     """Flash v4: tile num_heads÷4 max_seqlen_q÷2 max_seqlen_kv÷2."""
@@ -3631,7 +3591,10 @@ def test_flash_v4_tile_H_Lq_Lk():
 
 
 @pytest.mark.skip(
-    reason="Unsupported: propagate_named_dims bug — num_heads layout dim has no loop vars after view+transpose"
+    reason="Unsupported: partial reduction result consumed before"
+    " accumulation is complete (op reads a per-tile partial result from"
+    " the same loop group) -- same carry-propagation gap as the Lk"
+    " reduction-dim tiling cases below, not the view+transpose bug"
 )
 def test_flash_v4_tile_all():
     """Flash v4: tile all dims. B=2, H÷4, Lq÷2, Lk÷2."""
@@ -4380,33 +4343,14 @@ class TestCoarseTileSpyreHints(InductorTestCase):
         )
 
     # Consider deleting — superseded by Group 10 structured tests (_flash_v2_fn)
-    @pytest.mark.skip(
-        reason="finalize_layouts: restickify infeasible for copy ops across loop groups"
-    )
     def test_hint_flash_attention_v2(self):
         """Flash attention tiled over H (4 slices) via nested spyre_hints.
 
         Variant of test_hint_flash_attention with a causal mask and an
         explicit running-max (real_max) formulation that updates output and
-        denominator in place via copy_.
-
-        Still xfailed.  The divide sits outside the tiled scopes, so
-        `output`/`denominator` get full buffers + copy ops; each copy writes its
-        target without reading it, nothing costs that pairing, and
-        finalize_layouts overwrites the target with the writer's layout ->
-        "restickify needed but infeasible for op='buf24' input='buf26'".
-
-        Not resolvable by layout choice: writer and consumer need mutually
-        unrestickifiable candidates (forcing either aborts or gives ~70% wrong).
-        {"Lq": 2} alone reproduces it.  See
-        test_hint_flash_attention_v2_divide_in_scope for the formulation that
-        works, which localizes this to the cross-loop-group copy path.
-
-        Decision xfail: failing in CI (Actions run 30385154736, job
-        90362755639) on PR #3293. We've decided to xfail the coarse tiling
-        tests to allow us to merge to main -- deliberate decision to unblock
-        the merge, not a claim about a specific bisected root cause. Un-xfail
-        once the underlying regression is investigated and fixed.
+        denominator in place via copy_. The divide sits outside the tiled
+        scopes; see test_hint_flash_attention_v2_divide_in_scope for the
+        variant with the divide inside scope.
         """
         import math
         from torch_spyre._inductor import spyre_hint
@@ -4503,7 +4447,9 @@ class TestCoarseTileSpyreHints(InductorTestCase):
         )
 
     @pytest.mark.skip(
-        reason="Expected FixedTiledLayout for output buf — layout not promoted correctly with divide inside scope"
+        reason="Compiles now, but produces a genuine numeric inf at runtime"
+        " (99.3% of elements mismatched, abs diff inf) -- distinct from the"
+        " compile-time layout-promotion issue this test previously hit"
     )
     def test_hint_flash_attention_v2_divide_in_scope(self):
         """test_hint_flash_attention_v2 with the final divide INSIDE the scope.
@@ -4614,7 +4560,6 @@ class TestCoarseTileSpyreHints(InductorTestCase):
         )
 
     # Consider deleting — superseded by Group 10 structured tests (_flash_v3_fn)
-    @pytest.mark.skip(reason="dxp_standalone timeout")
     def test_hint_flash_attention_v3(self):
         from torch_spyre._inductor import spyre_hint
 
@@ -4720,7 +4665,10 @@ class TestCoarseTileSpyreHints(InductorTestCase):
         )
 
     @pytest.mark.skip(
-        reason="finalize_layouts: restickify infeasible for copy ops across loop groups"
+        reason="work_division: cannot satisfy span limit for a coarse-tile"
+        " read-copy tensor (coord evaluates to 4 after splits, best span"
+        " 8589934592 vs. limit 268435456) -- distinct from the previous"
+        " restickify-infeasibility failure this test hit"
     )
     def test_hint_flash_attention_v3_b2(self):
         """Same as flash_v3 but with B=2 and b_block_size=2 so B is nto tiled"""
@@ -4887,12 +4835,9 @@ class TestCoarseTileSpyreHints(InductorTestCase):
         )
 
     @pytest.mark.skip(
-        reason="propagate_named_dims bug: view+transpose produces index with var in two Mod "
-        "expressions that compute_coordinates cannot handle. "
-        "Root cause: find_repeat_vars skips len(mods)!=1 case silently; "
-        "compute_coordinates then produces coord=0 for num_heads dim. "
-        "Error (with PR#3034 fix): variable d2 (range 8192) appears in multiple Mod "
-        "expressions [Mod((d2//256), 32), Mod(d2, 256)] and cannot be mapped to coordinates."
+        reason="AssertionError in _stick_symbol: within-stick coordinate"
+        " carries 2 free symbols, want exactly 1 -- v4's view+transpose"
+        " layout produces a coordinate expression padding.py cannot solve"
     )
     def test_hint_flash_attention_v4(self):
         """This test attempts to replicate the standalone test_granite_attn.py with views
@@ -6039,231 +5984,6 @@ class TestCoarseTileSpyreHints(InductorTestCase):
             compare_with_cpu(
                 fn, x, y, run_compile=True, run_eager=False, atol=0.01, rtol=0.01
             )
-
-    # ------------------------------------------------------------------
-    # Per-bundle hbm_pool_sizes threading into SpyreKernel
-    # ------------------------------------------------------------------
-
-    @config.patch({"lx_planning": False})
-    def test_bundle_pool_size_threaded_from_hbm_pool_sizes(self):
-        """codegen_node must look up this bundle's own pool_size from
-        V.graph.hbm_pool_sizes, not a stale graph-global scalar.
-
-        lx_planning is disabled here so the `a = x + y` intermediate isn't
-        claimed by LX scratchpad planning first -- with LX planning on,
-        `add`/`mul`/`sub` outputs are all LX-eligible by default (see
-        OP_OUTPUT_NOT_GOOD_FOR_LX_REUSE in scratchpad/utils.py) and may win the
-        scratchpad before hbm_pool_planning ever sees them, leaving every
-        bundle's pool_size at 0 and proving nothing about the plumbing this
-        test exists to check.
-        """
-        from unittest.mock import patch
-
-        from torch_spyre._inductor.spyre_kernel import SpyreKernel
-
-        seen_pool_sizes = []
-        orig_init = SpyreKernel.__init__
-
-        def _recording_init(self, pool_size=0, **kwargs):
-            seen_pool_sizes.append(pool_size)
-            orig_init(self, pool_size=pool_size, **kwargs)
-
-        def fn(x, y):
-            a = x + y
-            b = a * 2
-            return b - x
-
-        x = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-        y = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-
-        with (
-            patch.object(SpyreKernel, "__init__", _recording_init),
-            mock_patch(_LAUNCH_JOBPLAN),
-            mock_patch(_PREPARE_KERNEL),
-            mock_patch("subprocess.run"),
-        ):
-            torch.compile(fn)(x, y)
-
-        self.assertTrue(seen_pool_sizes)
-        self.assertTrue(
-            any(seen_pool_sizes),
-            f"expected at least one bundle with a nonzero pool_size, got "
-            f"{seen_pool_sizes}",
-        )
-
-    @config.patch({"lx_planning": False})
-    @pytest.mark.filterwarnings("ignore::torch_spyre.ops.fallbacks.FallbackWarning")
-    def test_pool_alloc_scoped_per_bundle_across_fallback_boundary(self):
-        """A CPU-fallback op (torch.sin) splits the graph into multiple
-        bundles. With frontend_pool_allocation at its default (False), each
-        bundle's pool (if any) is allocated inside that bundle's own
-        generated MLIR via sdscbundle.device_mem_allocate -- there is no
-        Python-side pool tensor, and no bundle's MLIR references another
-        bundle's pool."""
-        from torch_spyre.execution import async_compile as async_compile_mod
-
-        def fn(t):
-            a = torch.exp(t) * 2  # compiled bundle 1; `a` crosses the
-            b = torch.sin(a)  # fallback op -- forces a bundle boundary
-            c = torch.exp(b) * 2  # compiled bundle 2
-            return c
-
-        x = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-
-        # get_output_dir() mints a fresh random tempdir on every call (see
-        # its uuid4()-based implementation), so it cannot be re-invoked from
-        # the test to recover the directory sdsc() actually used to write
-        # bundle.mlir. Wrap it to record kernel_name -> output_dir while
-        # still delegating to the real implementation.
-        real_get_output_dir = async_compile_mod.get_output_dir
-        output_dirs_by_kernel = {}
-
-        def _recording_get_output_dir(kernel_name):
-            output_dir = real_get_output_dir(kernel_name)
-            output_dirs_by_kernel[kernel_name] = output_dir
-            return output_dir
-
-        with (
-            mock_patch(_LAUNCH_JOBPLAN),
-            mock_patch(_PREPARE_KERNEL),
-            mock_patch("subprocess.run"),
-            mock_patch.object(
-                async_compile_mod, "get_output_dir", _recording_get_output_dir
-            ),
-            pytest.warns(UserWarning),
-        ):
-            _, source_codes = run_and_get_code(torch.compile(fn), x)
-        src = source_codes[0]
-
-        # No Python-side pool tensor of any kind remains in the wrapper.
-        # Ordinary output buffers legitimately use spyre_empty_with_layout,
-        # so distinguish pool allocs by their telltale uint8 dtype, same as
-        # test_no_python_side_pool_tensor_allocated below.
-        pool_alloc_lines = [
-            line
-            for line in src.splitlines()
-            if "spyre_empty_with_layout" in line and "uint8" in line
-        ]
-        self.assertEqual(pool_alloc_lines, [])
-        self.assertNotIn("_pool_", src)
-
-        # Each kernel name mentioned in an async_compile.sdsc(...) call gets
-        # its own bundle.mlir; any that uses a pool must self-allocate it via
-        # device_mem_allocate, never via a %pool_base_addr parameter.
-        kernel_names = re.findall(r"async_compile\.sdsc\('(\w+)'", src)
-        self.assertTrue(kernel_names)
-        saw_pool_allocate = False
-        for kernel_name in kernel_names:
-            self.assertIn(kernel_name, output_dirs_by_kernel)
-            bundle_path = os.path.join(
-                output_dirs_by_kernel[kernel_name], "bundle.mlir"
-            )
-            with open(bundle_path) as f:
-                bundle_text = f.read()
-            self.assertNotIn("%pool_base_addr", bundle_text)
-            if "device_mem_allocate" in bundle_text:
-                saw_pool_allocate = True
-        self.assertTrue(
-            saw_pool_allocate,
-            f"expected at least one bundle to use device_mem_allocate, "
-            f"kernels were {kernel_names}",
-        )
-
-    @config.patch({"lx_planning": False})
-    @pytest.mark.filterwarnings("ignore::torch_spyre.ops.fallbacks.FallbackWarning")
-    def test_no_python_side_pool_tensor_allocated(self):
-        """With frontend_pool_allocation at its default (False), no bundle
-        allocates a Python-side _pool_<name> tensor -- pool allocation is
-        entirely inside the generated MLIR."""
-
-        def fn(t):
-            a = torch.exp(t) * 2
-            b = torch.sin(a)  # fallback op -- forces a bundle boundary
-            c = torch.exp(b) * 2
-            return c
-
-        x = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-
-        with (
-            mock_patch(_LAUNCH_JOBPLAN),
-            mock_patch(_PREPARE_KERNEL),
-            mock_patch("subprocess.run"),
-            pytest.warns(UserWarning),
-        ):
-            _, source_codes = run_and_get_code(torch.compile(fn), x)
-        src = source_codes[0]
-
-        # No Python-side pool tensor allocation (uint8 SpyreTensorLayout) --
-        # ordinary output buffers still legitimately use
-        # spyre_empty_with_layout, so distinguish pool allocs by their
-        # telltale uint8 dtype, same as
-        # test_pool_alloc_scoped_per_bundle_across_fallback_boundary above.
-        pool_alloc_lines = [
-            line
-            for line in src.splitlines()
-            if "spyre_empty_with_layout" in line and "uint8" in line
-        ]
-        self.assertEqual(pool_alloc_lines, [])
-        self.assertNotIn("_pool_", src)
-
-    @config.patch({"lx_planning": False})
-    def test_pool_size_kwarg_in_generated_sdsc_call(self):
-        """define_kernel() must append pool_size=<N> to the generated
-        async_compile.sdsc(...) call text for a kernel whose pool_size > 0,
-        and pool_size=0 must never be emitted explicitly. See
-        test_pool_size_kwarg_omitted_when_no_pool below for the omission
-        case on a kernel with no pool usage at all."""
-
-        def fn(x, y):
-            a = x + y
-            b = a * 2
-            return b - x
-
-        x = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-        y = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-
-        with (
-            mock_patch(_LAUNCH_JOBPLAN),
-            mock_patch(_PREPARE_KERNEL),
-            mock_patch("subprocess.run"),
-        ):
-            _, source_codes = run_and_get_code(torch.compile(fn), x, y)
-        src = source_codes[0]
-
-        self.assertIn("pool_size=", src)
-        # Every async_compile.sdsc( call either has no pool_size kwarg, or a
-        # positive one -- pool_size=0 must never be emitted explicitly.
-        self.assertNotIn("pool_size=0", src)
-
-    @config.patch({"lx_planning": True})
-    def test_pool_size_kwarg_omitted_when_no_pool(self):
-        """A kernel with no pool usage gets no pool_size kwarg at all.
-
-        With lx_planning enabled, the `a = x + y` intermediate is claimed by
-        LX scratchpad planning before hbm_pool_planning ever sees it (see
-        OP_OUTPUT_NOT_GOOD_FOR_LX_REUSE in scratchpad/utils.py), so this bundle
-        has no pool-eligible buffer and define_kernel() must omit the
-        pool_size kwarg entirely rather than emit pool_size=0.
-        """
-
-        def fn(x, y):
-            a = x + y
-            b = a * 2
-            return b - x
-
-        x = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-        y = torch.randn(64, 64, dtype=torch.float16, device="spyre")
-
-        with (
-            mock_patch(_LAUNCH_JOBPLAN),
-            mock_patch(_PREPARE_KERNEL),
-            mock_patch("subprocess.run"),
-        ):
-            _, source_codes = run_and_get_code(torch.compile(fn), x, y)
-        src = source_codes[0]
-
-        self.assertIn("async_compile.sdsc(", src)
-        self.assertNotIn("pool_size", src)
 
 
 class TestNamedDimsHint(InductorTestCase):
